Create README.md
This was a test I received from a company for a front-end first stage. I was very happy that in the first 2 months of studies I received this request. It is already my third selection process at the moment.

The objective of this project was to develop a website based on the psd and the image LayoutTeste.jpg